Commit skeleton
Commit skeleton (30/11/2022)

+++ backend/src/database/model/product.model.ts
@@ -17,10 +17,10 @@ export class Product extends Model<Product> {
     @Column
     description: string
 
-    @Column(DataType.BIGINT)
+    @Column(DataType.INTEGER)
     price: number
 
-    @Column(DataType.BIGINT)
+    @Column(DataType.INTEGER)
     quantity_in_stock: number
     
     @Column(DataType.DATE(3))
